refactor: put game logics in it's own separate file
re #36
